Add files via upload
The user draws a box around the orange balls in the picture (roipoly). orange_pixels gathers the RGB vectors of the boxed area of the picture and stores it in an nx3 matrix. The mean of each column and the covariance matrix is calculated for p(x|orange). p(x|orange) is calculated using the Gaussian formula. 

This program isn't complete for testing.